pyproject.toml
[project]
name = "genai-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "openai>=2.14.0",
    "pathlib>=1.0.1",
    "playsound==1.2.2",
    "pydantic>=2.12.5",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
